IPv6 header

+++ src/crt/string.c
@@ -5,7 +5,7 @@
 
 char* strdup(const char* string)
 {
-	unsigned int len = strlen(string) * sizeof(char);
+	size_t len = strlen(string) * sizeof(char);
 	char* dup = (char*) malloc(len);
 	memcpy(dup, string, len);
 	return dup;
